requirements.txt
# ==============================
# Core Backend
# ==============================
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ==============================
# Machine Learning / LLM
# ==============================
torch>=2.2.0
transformers>=4.41.0
sentence-transformers>=2.7.0
accelerate>=0.30.0

# ==============================
# Vector Search
# ==============================
faiss-cpu>=1.8.0

# ==============================
# PDF Processing
# ==============================
pymupdf>=1.24.4

# ==============================
# Image OCR
# ==============================
pillow>=10.3.0
pytesseract>=0.3.10

# ==============================
# Language Detection & Translation
# ==============================
langdetect>=1.0.9
sacremoses>=0.1.1

# ==============================
# Video Processing (API-Free)
# ==============================
openai-whisper>=20231117
yt-dlp>=2024.5.27

# ==============================
# Utility
# ==============================
numpy>=1.26.0
scikit-learn>=1.4.2
tqdm>=4.66.4
python-dotenv>=1.0.1

# ==============================
# Optional (Production Improvements)
# ==============================
redis>=5.0.4
loguru>=0.7.2
